Exchange Rate Client

+++ exchange-rate/src/net/nanopay/exchangerate/model/ExchangeRate.js
@@ -22,7 +22,7 @@ foam.CLASS({
       name: 'toCurrency'
     },
     {
-      class: 'Long',
+      class: 'Double',
       name: 'rate'
     },
     {
